+++ src/main/kotlin/dev/aaronhowser/mods/aaron/registry/AaronMobEffectsRegistry.kt
@@ -1,6 +1,7 @@
 package dev.aaronhowser.mods.aaron.registry
 
 import net.minecraft.world.effect.MobEffect
+import net.minecraft.world.effect.MobEffectCategory
 import net.neoforged.neoforge.registries.DeferredHolder
 import net.neoforged.neoforge.registries.DeferredRegister
 import java.util.function.Supplier
@@ -11,6 +12,26 @@ abstract class AaronMobEffectsRegistry {
 
 	fun <T : MobEffect> register(id: String, supplier: Supplier<T>): DeferredHolder<MobEffect, T> {
 		return getMobEffectRegistry().register(id, supplier)
+	}
+
+	fun registerSimple(
+		id: String,
+		category: MobEffectCategory,
+		color: Int
+	): DeferredHolder<MobEffect, out MobEffect> {
+		return register(id) { object : MobEffect(category, color) {} }
+	}
+
+	fun registerSimpleInstantaneous(
+		id: String,
+		category: MobEffectCategory,
+		color: Int
+	): DeferredHolder<MobEffect, MobEffect> {
+		return register(id) {
+			object : MobEffect(category, color) {
+				override fun isInstantenous(): Boolean = true
+			}
+		}
 	}
 
 }